Make Explicit ldply is from plyr
plyr is loaded via namespace, so the ldply function is loaded, but not exposed to the user. ldply won't be found when the user runs the example unless they explicitly refer to the package with ::. In contrast rbind.fill is used within the queryMore function without any need for :: because of this.