Expand Test Coverage
Add unit tests for the REST API, Bulk API, and utilities of the package. The utilities are technically REST API calls, but do not involve changes in Salesforce records, so we split the testing scripts.

Fix rforcecom.search from erroring when there is a malformed search string. The new version handles malformed search strings and also returns a NULL object whenever no search results are found, which is more intuitive than a 1 row 0 column data.frame