Add External Id Field to Upsert Operation
Add the ability to specify an External Id field name which is required for performing the Bulk API Upsert Operation. The upserted dataset must have a column name that matches the External Id field specified when the job was created and the field must be configured in Salesforce as a External Id field.